==> ipynb notebooks codevita/01)Scraping CodeVita Results 2k16.ipynb <==
import pandas as pd
from six.moves import urllib
%qtconsole

from bs4 import BeautifulSoup

url=r'http://www.programminggeek.in/2016/08/tcs-codevita-2016-round-1-result.html'
csv_save_path=r'/mnt/ECFC8291FC8255AC/Datasets/Codevita Results/Codevita_2k16_round1_scrapped.csv'

html_pg=urllib.request.urlopen(url).read()

soup_wrapper=BeautifulSoup(html_pg)

soup_wrapper_table_body=soup_wrapper.find(id='post-body-1057866455376367824').tbody

soup_wrapper_table_body==None

soup_wrapper_table_body.prettify()[:500]

#Initializing lists for each columns
ranks=[]
team_name=[]
user_name=[]
tcs_region=[]
college_name=[]


for row in soup_wrapper_table_body.findAll('tr'):
    cells=row.findAll('td')
    if len(cells)==5: #Sanity Check
        ranks.append(cells[0].find(text=True))
        team_name.append(cells[1].find(text=True))
        user_name.append(cells[2].find(text=True))
        tcs_region.append(cells[3].find(text=True))
        college_name.append(cells[4].find(text=True))
    else:
        raise IndexError(cells)

[len(ranks),len(team_name),len(user_name),len(tcs_region),len(college_name)]

df_table=pd.DataFrame({'ranks':ranks,'team_name':team_name,'user_name':user_name,'tcs_region':tcs_region,'college_name':college_name})

df_table.describe()

Missing values in ranks and college name...

df_table[pd.isnull(df_table.college_name)]

df_table[pd.isnull(df_table.ranks)]

On checking the original url, it can be seen that this particular row has the rank wrongly embedded in the team name column...and also a part of the username is embedded in the team name column

df_table.ranks.loc[3377]=3378

df_table.team_name.loc[3377]='Tweak It'

df_table.user_name.loc[3377]='Vijay Saayi R & Siva P'

df_table.loc[3377]

pd.set_option('display.max_rows',200)
df_table.tcs_region.value_counts()

df_table.to_csv(csv_save_path)

df_table[df_table.tcs_region=='Haldia']
