==> ipynb notebooks codevita/02)Cleaning 2k15 TechVita Results Part 1.ipynb <==
#%qtconsole
import pandas as pd
import platform

file_name='codevita 2k15 scrapped.csv'
def return_file_path(base_url,file_name,file_seperator,intermediate_folder=''):
    if not intermediate_folder=='':
        intermediate_folder=intermediate_folder+file_seperator
    return '{}{}{}'.format(base_url,intermediate_folder,file_name)

if platform.system()=='Linux':
    base_url=r'/mnt/ECFC8291FC8255AC/Datasets/Codevita Results/'
    file_seperator='/'
else:
    file_seperator='\\'
    base_url='F:\\Datasets\\Codevita Results\\'
    
    

if platform.system()=='Linux':
    results_2k15_df=pd.read_csv(return_file_path(base_url,file_name,file_seperator))
else:
    results_2k15_df=pd.read_csv(return_file_path(base_url,file_name,file_seperator))

results_2k15_df.shape

results_2k15_df.describe()

results_2k15_df.head()

Removing all the spaces from the names of the columns in the DF for easier access as a property of the DF object instead of having to index into a specific column

results_2k15_df['team_name']=results_2k15_df['Team Name']
results_2k15_df=results_2k15_df.drop('Team Name',axis=1)

results_2k15_df['tcs_region']=results_2k15_df['TCS Region']
results_2k15_df=results_2k15_df.drop('TCS Region',axis=1)

results_2k15_df['users_name']=results_2k15_df['Users Name']
results_2k15_df=results_2k15_df.drop('Users Name',axis=1)

results_2k15_df['college_name']=results_2k15_df['College Name']
results_2k15_df=results_2k15_df.drop('College Name',axis=1)

results_2k15_df.head()

Checking whether each entry of the users_name column has a '&'...if it has two entries, only then will the '&' be present otherwise it will not be present...if not present it will raise an assertion error

assert(sum(results_2k15_df.users_name.apply(lambda x:not '&' in x))==0)

results_2k15_df['user_1']=results_2k15_df.users_name.apply(lambda x:x.split(sep='&')[0].strip())
results_2k15_df['user_2']=results_2k15_df.users_name.apply(lambda x:x.split(sep='&')[1].strip())

Checking for any null entries in user_1 or user_2

assert(results_2k15_df.user_1[results_2k15_df.user_1==''].shape[0]==0)
assert(results_2k15_df.user_2[results_2k15_df.user_2==''].shape[0]==0)

Checking whether the column of tcs_regions contains any invalid entries

results_2k15_df.tcs_region.unique()

Observation: Many colleges have the name of the department attached before them...all of those start with the word 'Dept'

results_2k15_df.college_name[results_2k15_df.college_name.str.startswith('Dept')]

results_2k15_df.college_name[results_2k15_df.college_name.str.startswith('Dept')].unique()

Extracting the name of the college from the above results...note that we can extract the results by splitting it at the occurence of the first comma

results_2k15_df.college_name=results_2k15_df.college_name.apply(lambda x:x.split(sep=',')[1] if x.startswith('Dept') else x)

results_2k15_df.college_name[results_2k15_df.college_name.str.startswith('Dept')]

Checking for empty values in team_name

results_2k15_df.team_name[results_2k15_df.team_name=='']

college_df=pd.DataFrame(results_2k15_df,columns=['college_name','tcs_region'])

college_df.shape

college_df.head()

college_df=college_df.drop_duplicates()
college_df.college_name=college_df.college_name.str.strip()

college_df.to_csv(return_file_path(base_url,'college_list_2k15.csv',file_seperator,intermediate_folder='generated_preprocessing'))

results_2k15_df.to_csv(return_file_path(base_url,'codevita_2k15_preprocessed.csv',file_seperator,intermediate_folder='generated_preprocessing'))
