==> ipynb notebooks codevita/03)Preprocessing 2k16 TechVita Results.ipynb <==
#%qtconsole #To open a seperate window running the Jupyter Qt Console which is running the same kernel
import pandas as pd
import platform
import numpy as np

file_name='Codevita_2k16_round1_scrapped.csv'
def return_file_path(base_url,file_name,file_seperator,intermediate_folder=''):
    if not intermediate_folder=='':
        intermediate_folder=intermediate_folder+file_seperator
    return '{}{}{}'.format(base_url,intermediate_folder,file_name)

if platform.system()=='Linux':
    base_url=r'/mnt/ECFC8291FC8255AC/Datasets/Codevita Results/'
    file_seperator='/'
else:
    file_seperator='\\'
    base_url='F:\\Datasets\\Codevita Results\\'
    
    

if platform.system()=='Linux':
    results_2k16_df=pd.read_csv(return_file_path(base_url,file_name,file_seperator))
else:
    results_2k16_df=pd.read_csv(return_file_path(base_url,file_name,file_seperator))

results_2k16_df.head()

results_2k16_df=results_2k16_df.drop('Unnamed: 0',axis=1)

results_2k16_df.shape

results_2k16_df.describe()

results_2k16_df.head()

Trimming each string value to ensure there are no unnecessary whitespaces

results_2k16_df.college_name=results_2k16_df.college_name.str.strip()
results_2k16_df.tcs_region=results_2k16_df.tcs_region.str.strip()
results_2k16_df.team_name=results_2k16_df.team_name.str.strip()
results_2k16_df.user_name=results_2k16_df.user_name.str.strip()

Checking whether each entry of the users_name column has a '&'...if it has two entries, only then will the '&' be present otherwise it will not be present...if not present it will raise an assertion error

assert(sum(results_2k16_df.user_name.apply(lambda x:not '&' in x))==0)

results_2k16_df['user_1']=results_2k16_df.user_name.apply(lambda x:x.split(sep='&')[0].strip())
results_2k16_df['user_2']=results_2k16_df.user_name.apply(lambda x:x.split(sep='&')[1].strip())

Checking for any null entries in user_1 or user_2

assert(results_2k16_df.user_1[results_2k16_df.user_1==''].shape[0]==0)
assert(results_2k16_df.user_2[results_2k16_df.user_2==''].shape[0]==0)

Checking whether the column of tcs_regions contains any invalid entries

results_2k16_df.tcs_region.unique()

We can see above that most probably due to errors induced during scraping the results of the website,a lot of erronous values got into the tcs_region column

On manual inspection, it was observed that for those *tcs_region* cells which contains erronous values, if we append the corresponding value of the *college_name* to that of the tcs_region, we get the actual name of the colleges 

Also, it was observed that many of the colleges are represented by more than 1 names in the *college_name* column...for eg. *'Srm Engineering College., Kattankulathur'* and *'SRM University'* are both same yet they are referred to by different times

To deal with this problem, let us first make a list of those college names which we know for a fact is a true region and not an erronous value...the easiest way for us to do this is by getting a frequency count of each of the entries in the tcs_region columns...the one which is a true region should have much more frequency count 

results_2k16_df.tcs_region.value_counts()

From above, it is clear that only the top 10 most frequently present values in the *tcs_region* column are actual regions while the others are erronous entries

true_tcs_regions_2k16=results_2k16_df.tcs_region.value_counts().iloc[0:10]
erronous_tcs_regions_2k16=results_2k16_df.tcs_region.value_counts().iloc[10:]

Now for the erronous entries, let's append the name of the colleges to the erronous entries in the *tcs_region* to create the modified college name

true_tcs_regions_index=true_tcs_regions_2k16.index
results_2k16_df['college_name_modified']=results_2k16_df.apply(lambda x:'{} {}'.format(x['tcs_region'],x['college_name']) if x['tcs_region'] not in true_tcs_regions_index else x['college_name'],axis=1)

results_2k16_df.head()

Doing a quick check by checking the values of some of the erring examples

results_2k16_df[(results_2k16_df.tcs_region=='V.S.B.') | (results_2k16_df.tcs_region=='DEENBANDHU')]

Now, we observe that for those entries which have erronous values in the *tcs_region* section, we have the actual region as the last word in the user_name column and therefore the last name in the user_2 column

def correct_tcs_region_column(x):
    if not x['tcs_region'] in true_tcs_regions_2k16:
        true_region=x['user_2'].split()[-1:][0]
        assert(true_region in true_tcs_regions_2k16) #Assuring whether the extracted region is present in the list of true regions
        return true_region
    else:
        return x['tcs_region']

results_2k16_df['tcs_region_modified']=results_2k16_df.apply(correct_tcs_region_column,axis=1)

def correct_user2_column(x):
    if not x['tcs_region'] in true_tcs_regions_2k16:
        modified_username=' '.join(x['user_2'].split()[:-1])
        return modified_username
    else:
        return x['user_2']

results_2k16_df['user_2']=results_2k16_df.apply(correct_user2_column,axis=1)

results_2k16_df.head()

results_2k16_df[(results_2k16_df.tcs_region=='V.S.B.') | (results_2k16_df.tcs_region=='DEENBANDHU')].head()

results_2k16_df.tcs_region_modified.value_counts()

We had seen while preprocessing the 2k15 results that many colleges have the name of the department attached before them...all of those start with the word 'Dept'

print(results_2k16_df.college_name_modified.count())
results_2k16_df.college_name_modified.dropna().count()


results_2k16_df.team_name[results_2k16_df.team_name=='']

college_db=pd.read_csv(return_file_path(base_url,'college_list.csv',file_seperator,'generated_preprocessing'))

college_db.head()

college_db=college_db.drop('Unnamed: 0',axis=1)

college_db.head()

college_db.shape

len(results_2k16_df.college_name_modified.unique())

Checking whether the regions of 2k15 and 2k16 are the same or not

pd.Series(results_2k16_df.tcs_region_modified.unique()).isin(college_db.tcs_region.unique())

results_2k16_df.head()

college_db_2k16=pd.DataFrame(results_2k16_df,columns=['college_name','college_name_modified','tcs_region','tcs_region_modified'])

Since we are creating a db of the colleges present in the 2k16 results, we need to remove all the duplicate entries present in the *college_name_modified* column

college_db_2k16=college_db_2k16.ix[college_db_2k16.college_name_modified.drop_duplicates().index]

college_db_2k16.to_csv(return_file_path(base_url,'college_list_2k16.csv',file_seperator,'generated_preprocessing'))

results_2k16_df.to_csv(return_file_path(base_url,'results_2k16_preprocessed.csv',file_seperator,'generated_preprocessing'),index=False)
