==> ipynb notebooks codevita/05) v2 Trying to match colleges better from 2k15-2k16 list to 2k17.ipynb <==
import pandas as pd
import numpy as np
#%qtconsole
import platform
import matplotlib.pyplot as plt
%matplotlib inline

def return_file_path(base_url,file_name,file_seperator,intermediate_folder=''):
    if not intermediate_folder=='':
        intermediate_folder=intermediate_folder+file_seperator
    return '{}{}{}'.format(base_url,intermediate_folder,file_name)

if platform.system()=='Linux':
    base_url=r'/mnt/ECFC8291FC8255AC/Datasets/Codevita Results/'
    file_seperator='/'
else:
    file_seperator='\\'
    base_url='F:\\Datasets\\Codevita Results\\'
intermediate_folder=''

college_prev=pd.read_excel('./college_combined_2k15_2k16.xlsx')
college_prev=college_prev.rename(columns={'college_name_modified_normalized':'college_name_normalized',
                                          'coll_name_mod_norm_wo_common_words':'coll_name_norm_wo_common_words',
                                         'coll_name_mod_norm_wo_common_words_v2':'coll_name_norm_wo_common_words_v2'})

college_prev.head()

college_2k17=pd.read_csv(return_file_path(base_url,'codevita_2k17_scrapped.csv',file_seperator))

college_2k17.head()

print(college_2k17.shape)
college_2k17.notnull().sum()

college_2k17=college_2k17.rename(columns={'zonal_name':'tcs_region'})

college_2k17_list=pd.DataFrame(college_2k17,columns=['college_name','tcs_region'])
college_2k17_list=college_2k17_list
college_2k17_list.college_name.unique().shape[0]

college_2k17_list=college_2k17_list.groupby('college_name',as_index=False).first()
print(college_2k17_list.shape)
college_2k17_list.head()

college_2k17_list['index']=college_2k17_list.index
college_2k17_list['mod_index']=college_2k17_list['index'].apply(lambda x:'2k17_'+str(x))
college_2k17_list=college_2k17_list.drop('index',axis=1)

college_2k17_list.head()

Using string matching normalized to check the common entries between colleges in 2k15 and 2k16

from nltk.tokenize import RegexpTokenizer
tokenizer = RegexpTokenizer(r'\w+')

def modify_college_names_apply(x):
    return ' '.join(tokenizer.tokenize(x.lower()))
college_2k17_list['college_name_normalized']=college_2k17_list.college_name.apply(modify_college_names_apply)

college_2k17_list.head()

print(college_2k17_list.shape)
college_2k17_list.notnull().sum() #Ensuring that the newly formed column doesnt have any null values

print(sum(college_2k17_list.college_name_normalized.isin(college_prev.college_name_normalized)))
366/499*100

college_list_normalized_matches=pd.merge(college_prev,college_2k17_list,on='college_name_normalized',suffixes=('_prev','_2k17'))

college_list_normalized_matches.head()

def find(dic,index):
    return dic[index]
def union(dic,index1,index2):
    root1=find(dic,index1)
    root2=find(dic,index2)
    for key,value in dic.items():
        if value==root2:
            #print('Changed key {} from {} to {}'.format(key,value,root1))
            dic[key]=root1

match_dict={}
for _,row in college_prev.iterrows():
    match_dict[row['mod_index']]=row['master_mod_index']
for _,row in college_2k17_list.iterrows():
    match_dict[row['mod_index']]=row['mod_index']
len(match_dict.keys())

for _,row in college_list_normalized_matches.iterrows():
    union(match_dict,row['mod_index_prev'],row['mod_index_2k17'])

match_dict['2k17_24'] #Sanity Check

common_words=['of','college','technology','engineering','institute','university','engg',
              'and','inst','univ','management','group','for','information','school']
def modify_college_names_wo_common_words_apply(x):
    global common_words
    words=tokenizer.tokenize(x)
    modified_words=[]
    for word in words:
        if word not in common_words:
            modified_words.append(word)
    return ' '.join(modified_words)
college_2k17_list['coll_name_norm_wo_common_words']=college_2k17_list.college_name_normalized.apply(modify_college_names_wo_common_words_apply)

print(sum(college_2k17_list.coll_name_norm_wo_common_words==''))
college_2k17_list[college_2k17_list.coll_name_norm_wo_common_words=='']

college_2k17_list['coll_name_norm_wo_common_words_v2']=college_2k17_list.coll_name_norm_wo_common_words
college_2k17_list.loc[206,'coll_name_norm_wo_common_words_v2']='kolkata iem'

sum(college_2k17_list.coll_name_norm_wo_common_words_v2.isin(college_prev.coll_name_norm_wo_common_words_v2))

index_wo_common_words_matches=college_2k17_list[college_2k17_list.coll_name_norm_wo_common_words_v2.isin(college_prev.coll_name_norm_wo_common_words_v2)]['mod_index']
index_prev_matches_till_normalization=college_list_normalized_matches.mod_index_2k17.unique()
index_new_matches=np.setdiff1d(index_wo_common_words_matches,index_prev_matches_till_normalization)
mismatch_diff=np.setdiff1d(index_prev_matches_till_normalization,index_wo_common_words_matches)
index_new_matches

df_new_matches_2k17=college_2k17_list[college_2k17_list.mod_index.isin(index_new_matches)]
df_new_matches_wo_common_words_join=pd.merge(df_new_matches_2k17,college_prev,on='coll_name_norm_wo_common_words_v2',suffixes=('_2k17','_prev'))
print(df_new_matches_wo_common_words_join.shape)
df_new_matches_wo_common_words_join

df_new_matches_wo_common_words_join.to_excel('./2k17_new_matches_wo_common_words.xlsx',index=False)

df_new_matches_wo_common_words_matches=pd.read_excel('./2k17_new_matches_wo_common_words(copy).xlsx')
df_new_matches_wo_common_words_matches=df_new_matches_wo_common_words_matches[df_new_matches_wo_common_words_matches.match=='y']
print(df_new_matches_wo_common_words_matches.shape)
df_new_matches_wo_common_words_matches.head()

for _,row in df_new_matches_wo_common_words_matches.iterrows():
    union(match_dict,row['mod_index_prev'],row['mod_index_2k17'])
match_dict['2k17_83']=='2k16_628' #Sanity Check

#Doing the next step just to make the college_2k17_list fully compatible with the college_prev list
college_2k17_list['coll_name_case_independent']=college_2k17_list.college_name.apply(lambda x:x.lower())
college_2k17_list['master_mod_index']=college_2k17_list.mod_index.apply(lambda x:match_dict[x])

college_prev.shape

college_prev=college_prev.append(college_2k17_list,ignore_index=True)

college_prev.shape

college_prev.columns

college_prev.to_excel('./college_combined_2k15_2k16_2k17.xlsx',index=True)
