==> ipynb notebooks codevita/06) Linking each individual result with the master_mod_index.ipynb <==
import pandas as pd
import numpy as np
#%qtconsole
import platform
import matplotlib.pyplot as plt
%matplotlib inline

def return_file_path(base_url,file_name,file_seperator,intermediate_folder=''):
    if not intermediate_folder=='':
        intermediate_folder=intermediate_folder+file_seperator
    return '{}{}{}'.format(base_url,intermediate_folder,file_name)

if platform.system()=='Linux':
    base_url=r'/mnt/ECFC8291FC8255AC/Datasets/Codevita Results/'
    file_seperator='/'
else:
    file_seperator='\\'
    base_url='F:\\Datasets\\Codevita Results\\'
intermediate_folder=''

college_list=pd.read_excel('./college_combined_2k15_2k16_2k17.xlsx')
college_list.head()

college_list=college_list.sort_values('master_mod_index')

college_list.head()

results_2k15_df=pd.read_csv(return_file_path(base_url,'codevita_2k15_preprocessed.csv',file_seperator,'generated_preprocessing'))
results_2k16_df=pd.read_csv(return_file_path(base_url,'results_2k16_preprocessed.csv',file_seperator,'generated_preprocessing'))
results_2k17_df=pd.read_csv(return_file_path(base_url,'codevita_2k17_scrapped.csv',file_seperator))

(results_2k15_df.shape[0],results_2k16_df.shape[0],results_2k17_df.shape[0])

See screenshots 6 and 7 to understand why it is necessary to remove leading and trailing whitespaces

results_2k15_df.columns

results_2k16_df.columns

results_2k17_df.columns

results_2k15_df.college_name=results_2k15_df.college_name.str.strip()
results_2k16_df.college_name=results_2k16_df.college_name_modified.str.strip()
results_2k17_df.college_name=results_2k17_df.college_name.str.strip()

college_list.columns

#Slicing the college list year wise
college_list_2k15=college_list[college_list.mod_index.str.startswith('2k15_')]
college_list_2k16=college_list[college_list.mod_index.str.startswith('2k16_')]
college_list_2k17=college_list[college_list.mod_index.str.startswith('2k17_')]

assert college_list_2k17.shape[0]+college_list_2k16.shape[0]+college_list_2k15.shape[0]==college_list.shape[0]
print((college_list_2k15.shape[0],college_list_2k16.shape[0],college_list_2k17.shape[0]))

results_2k15_df=results_2k15_df.merge(college_list_2k15.loc[:,['mod_index','master_mod_index','college_name']],how='left',left_on='college_name',right_on='college_name')

results_2k15_df.head()

not_matched_mod_index_2k15=np.setdiff1d(college_list_2k15.mod_index.values,list(results_2k15_df.mod_index.unique()))
assert len(not_matched_mod_index_2k15)==0

results_2k16_df=results_2k16_df.rename(columns={'college_name':'college_name_prerectified','college_name_modified':'college_name','tcs_region':'tcs_region_prerectified','tcs_region_modified':'tcs_region'})

results_2k16_df=results_2k16_df.merge(college_list_2k16.loc[:,['mod_index','master_mod_index','college_name']],how='left',left_on='college_name',right_on='college_name')

results_2k16_df.head()

not_matched_mod_index_2k16=np.setdiff1d(college_list_2k16.mod_index.values,list(results_2k16_df.mod_index.unique()))
not_matched_mod_index_2k16
#assert len(not_matched_mod_index_2k16)==0

college_list_2k16[college_list_2k16.mod_index.isin(not_matched_mod_index_2k16)]

The single rogue entry below can be accepted since it is a nil value or an unknown value

results_2k17_df=results_2k17_df.merge(college_list_2k17.loc[:,['mod_index','master_mod_index','college_name']],how='left',left_on='college_name',right_on='college_name')

results_2k17_df.head()

not_matched_mod_index_2k17=np.setdiff1d(college_list_2k17.mod_index.values,list(results_2k17_df.mod_index.unique()))
not_matched_mod_index_2k17
assert len(not_matched_mod_index_2k17)==0

Linking master college names and master college regions

results_2k15_df=results_2k15_df.merge(college_list.loc[:,['college_name','tcs_region','mod_index']],left_on='master_mod_index',right_on='mod_index',suffixes=('','_master'))

results_2k15_df=results_2k15_df.drop('mod_index_master',axis=1)

results_2k15_df.head()

results_2k16_df=results_2k16_df.merge(college_list.loc[:,['college_name','tcs_region','mod_index']],left_on='master_mod_index',right_on='mod_index',suffixes=('','_master'))
results_2k16_df=results_2k16_df.drop('mod_index_master',axis=1)
results_2k16_df.head()

results_2k17_df=results_2k17_df.merge(college_list.loc[:,['college_name','tcs_region','mod_index']],left_on='master_mod_index',right_on='mod_index',suffixes=('','_master'))
results_2k17_df=results_2k17_df.drop('mod_index_master',axis=1)
results_2k17_df.head()

results_2k17_df=results_2k17_df.rename(columns={'tcs_region':'tcs_region_master'})

results_2k17_df[(results_2k17_df.tcs_region_master=='zone1') & (results_2k17_df.zonal_name=='zone3')]

results_2k15_df.to_excel('./results_2k15_backlinked.xlsx',index=False)
results_2k16_df.to_excel('./results_2k16_backlinked.xlsx',index=False)
results_2k17_df.to_excel('./results_2k17_backlinked.xlsx',index=False)
