==> ipynb notebooks codevita/07) Individual names database.ipynb <==
import pandas as pd
import numpy as np
#%qtconsole
import platform
import matplotlib.pyplot as plt
%matplotlib inline

def return_file_path(base_url,file_name,file_seperator,intermediate_folder=''):
    if not intermediate_folder=='':
        intermediate_folder=intermediate_folder+file_seperator
    return '{}{}{}'.format(base_url,intermediate_folder,file_name)

if platform.system()=='Linux':
    base_url=r'/mnt/ECFC8291FC8255AC/Datasets/Codevita Results/'
    file_seperator='/'
else:
    file_seperator='\\'
    base_url='F:\\Datasets\\Codevita Results\\'
intermediate_folder=''

results_2k15_df=pd.read_excel('./results_2k15_backlinked.xlsx')
results_2k16_df=pd.read_excel('./results_2k16_backlinked.xlsx')
results_2k17_df=pd.read_excel('./results_2k17_backlinked.xlsx')

results_2k15_df['year']='2k15'
results_2k16_df['year']='2k16'
results_2k17_df['year']='2k17'

results_2k15_df.columns

results_2k16_df.columns

results_2k17_df.columns

results_2k15_df.user_1=results_2k15_df.user_1.str.strip().str.lower()
results_2k15_df.user_2=results_2k15_df.user_2.str.strip().str.lower()

results_2k16_df.user_1=results_2k16_df.user_1.str.strip().str.lower()
results_2k16_df.user_2=results_2k16_df.user_2.str.strip().str.lower()

results_2k17_df.user_name=results_2k17_df.user_name.str.strip().str.lower()

from nltk.tokenize import RegexpTokenizer
tokenizer = RegexpTokenizer(r'\w+')

results_2k15_users=results_2k15_df.loc[:,['user_1','user_2','college_name_master', 'tcs_region_master',
       'mod_index','year','master_mod_index']]
results_2k16_users=results_2k16_df.loc[:,['user_1','user_2','college_name_master', 'tcs_region_master',
       'mod_index','year','master_mod_index']]
results_2k17_users=results_2k17_df.loc[:,['user_name','college_name_master', 'tcs_region_master',
       'mod_index','year','master_mod_index']].rename(columns={'user_name':'user'})

results_combined=pd.DataFrame(results_2k17_users,columns=['year','user','college_name_master','tcs_region_master','mod_index','master_mod_index'])

results_2k15_users_part1=results_2k15_df.loc[:,['year','user_1','college_name_master', 'tcs_region_master','mod_index',
       'master_mod_index']].rename(columns={'user_1':'user'})
results_2k15_users_part2=results_2k15_df.loc[:,['year','user_2','college_name_master', 'tcs_region_master','mod_index',
       'master_mod_index']].rename(columns={'user_2':'user'})
results_2k16_users_part1=results_2k16_df.loc[:,['year','user_1','college_name_master', 'tcs_region_master','mod_index',
       'master_mod_index']].rename(columns={'user_1':'user'})
results_2k16_users_part2=results_2k16_df.loc[:,['year','user_2','college_name_master', 'tcs_region_master','mod_index',
       'master_mod_index']].rename(columns={'user_2':'user'})

print((sum(results_2k15_users_part1.user.isnull()),sum(results_2k15_users_part2.user.isnull()),sum(results_2k16_users_part1.user.isnull()),sum(results_2k16_users_part2.user.isnull()),sum(results_2k17_users.user.isnull())))

results_combined=results_combined.append(results_2k15_users_part1,ignore_index=True)
results_combined=results_combined.append(results_2k15_users_part2,ignore_index=True)
results_combined=results_combined.append(results_2k16_users_part1,ignore_index=True)
results_combined=results_combined.append(results_2k16_users_part2,ignore_index=True)

results_combined.shape

sum(results_combined.user.isnull())

sum(results_combined.master_mod_index.isnull())

results_combined.columns

unique_users=results_combined.drop_duplicates(['user','master_mod_index'])

sum(unique_users.user.isnull())

sum(unique_users.master_mod_index.isnull())

unique_users.shape

unique_users.year.value_counts()

results_combined.shape[0]-unique_users.shape[0]

unique_users=unique_users.reindex(list(range(unique_users.shape[0])))

unique_users.tail()

unique_users=unique_users.loc[:,['year','user','college_name_master','tcs_region_master','master_mod_index']]

unique_users.shape

unique_users.to_excel('./2k15_2k16_2k17_unique_users.xlsx',index=False)
