==> ipynb notebooks codevita/08) Backlinking users to the results.ipynb <==
import pandas as pd
import numpy as np
#%qtconsole
import platform
import matplotlib.pyplot as plt
%matplotlib inline

def return_file_path(base_url,file_name,file_seperator,intermediate_folder=''):
    if not intermediate_folder=='':
        intermediate_folder=intermediate_folder+file_seperator
    return '{}{}{}'.format(base_url,intermediate_folder,file_name)

if platform.system()=='Linux':
    base_url=r'/mnt/ECFC8291FC8255AC/Datasets/Codevita Results/'
    file_seperator='/'
else:
    file_seperator='\\'
    base_url='F:\\Datasets\\Codevita Results\\'
intermediate_folder=''

Loading up the results list

results_2k15_df=pd.read_excel('./results_2k15_backlinked.xlsx')
results_2k16_df=pd.read_excel('./results_2k16_backlinked.xlsx')
results_2k17_df=pd.read_excel('./results_2k17_backlinked.xlsx')

college_list_df=pd.read_excel('./college_combined_2k15_2k16_2k17.xlsx')

users_list_df=pd.read_excel('./2k15_2k16_2k17_unique_users.xlsx')

#normalizing the user_names
results_2k17_df.user_name=results_2k17_df.user_name.str.strip().str.lower()
results_2k15_df.user_1=results_2k15_df.user_1.str.strip().str.lower()
results_2k15_df.user_2=results_2k15_df.user_2.str.strip().str.lower()
results_2k16_df.user_1=results_2k16_df.user_1.str.strip().str.lower()
results_2k16_df.user_2=results_2k16_df.user_2.str.strip().str.lower()

results_2k15_df.head()

results_2k16_df.head()

results_2k17_df.head()

college_list_df.head()

results_2k15_df['year']='2k15'
results_2k16_df['year']='2k16'
results_2k17_df['year']='2k17'

results_2k15_df.columns

results_2k16_df.columns

results_2k17_df.columns

users_list_df.columns

users_list_df.head()

users_list_df_size=users_list_df.shape[0]

users_list_df['id_user']=users_list_df.index

users_list_df.tail()

results_2k15_df_temp=results_2k15_df.merge(users_list_df.loc[:,['user','id_user','master_mod_index']],how='left',
                                           left_on=['user_1','master_mod_index'],right_on=['user','master_mod_index'],suffixes=('','_user1'))

results_2k15_df_temp.head()

results_2k15_df_temp.tail()

print(sum(results_2k15_df_temp.id_user.isnull()))
print(sum(results_2k15_df_temp.user.isnull()))

results_2k15_df_temp=results_2k15_df_temp.drop('user',axis=1)
results_2k15_df_temp.head()

results_2k15_df_temp=results_2k15_df_temp.rename(columns={'id_user':'id_user1'})

results_2k15_df=results_2k15_df_temp

results_2k15_df_temp=results_2k15_df.merge(users_list_df.loc[:,['user','id_user','master_mod_index']],how='left',
                                           left_on=['user_2','master_mod_index'],right_on=['user','master_mod_index'],suffixes=('','_user2'))

results_2k15_df_temp.head()

print(sum(results_2k15_df_temp.id_user.isnull()))
print(sum(results_2k15_df_temp.user.isnull()))

results_2k15_df_temp=results_2k15_df_temp.drop('user',axis=1)
results_2k15_df_temp=results_2k15_df_temp.rename(columns={'id_user':'id_user2'})
results_2k15_df=results_2k15_df_temp

results_2k15_df.head()

results_2k16_df_temp=results_2k16_df.merge(users_list_df.loc[:,['user','id_user','master_mod_index']],how='left',
                                           left_on=['user_1','master_mod_index'],right_on=['user','master_mod_index'],suffixes=('','_user1'))
results_2k16_df_temp.head()

print(sum(results_2k16_df_temp.id_user.isnull()))
print(sum(results_2k16_df_temp.user.isnull()))
results_2k16_df_temp=results_2k16_df_temp.drop('user',axis=1)
results_2k16_df_temp=results_2k16_df_temp.rename(columns={'id_user':'id_user1'})
results_2k16_df=results_2k16_df_temp
results_2k16_df.head()

results_2k16_df_temp=results_2k16_df.merge(users_list_df.loc[:,['user','id_user','master_mod_index']],how='left',
                                           left_on=['user_2','master_mod_index'],right_on=['user','master_mod_index'],suffixes=('','_user2'))
results_2k16_df_temp.head()

print(sum(results_2k16_df_temp.id_user.isnull()))
print(sum(results_2k16_df_temp.user.isnull()))
results_2k16_df_temp=results_2k16_df_temp.drop('user',axis=1)
results_2k16_df_temp=results_2k16_df_temp.rename(columns={'id_user':'id_user2'})
results_2k16_df=results_2k16_df_temp
results_2k16_df.head()

results_2k17_df_temp=results_2k17_df.merge(users_list_df.loc[:,['user','id_user','master_mod_index']],how='left',
                                           left_on=['user_name','master_mod_index'],right_on=['user','master_mod_index'],suffixes=('','_user1'))
results_2k17_df_temp.head()

print(sum(results_2k17_df_temp.id_user.isnull()))
print(sum(results_2k17_df_temp.user.isnull()))
results_2k17_df_temp=results_2k17_df_temp.drop('user',axis=1)
results_2k17_df_temp=results_2k17_df_temp.rename(columns={'id_user':'id_user1','user_name':'user1'})
results_2k17_df=results_2k17_df_temp
results_2k17_df.head()

results_2k15_df=results_2k15_df.drop('Unnamed: 0',axis=1)
results_2k17_df=results_2k17_df.drop('Unnamed: 0',axis=1)

results_2k15_df.to_excel('./results_2k15_df_backlinking_users.xlsx',index=False)
results_2k16_df.to_excel('./results_2k16_df_backlinking_users.xlsx',index=False)
results_2k17_df.to_excel('./results_2k17_df_backlinking_users.xlsx',index=False)

users_list_df.to_excel('./2k15_2k16_2k17_unique_users.xlsx',index=False)
