==> ipynb notebooks codevita/09) Preparing data for Tableau VIsualization.ipynb <==
import pandas as pd
import numpy as np
#%qtconsole
import platform
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

def return_file_path(base_url,file_name,file_seperator,intermediate_folder=''):
    if not intermediate_folder=='':
        intermediate_folder=intermediate_folder+file_seperator
    return '{}{}{}'.format(base_url,intermediate_folder,file_name)

if platform.system()=='Linux':
    base_url=r'/mnt/ECFC8291FC8255AC/Datasets/Codevita Results/'
    file_seperator='/'
else:
    file_seperator='\\'
    base_url='F:\\Datasets\\Codevita Results\\'
intermediate_folder=''

results_2k15_df=pd.read_excel('./results_2k15_df_backlinking_users.xlsx')
results_2k16_df=pd.read_excel('./results_2k16_df_backlinking_users.xlsx')
results_2k17_df=pd.read_excel('./results_2k17_df_backlinking_users.xlsx')

results_2k15_df.columns

results_2k16_df.columns

results_2k17_df.columns

(results_2k15_df.Rank.is_unique,results_2k16_df.ranks.is_unique)

#### Adding regional ranks for 2k15 and 2k16

regions=results_2k15_df.tcs_region.unique()
results_2k15_df=results_2k15_df.sort_values('Rank',ascending=True)
results_2k16_df=results_2k16_df.sort_values('ranks',ascending=True)
for region in regions:
    subset_mask_2k15=results_2k15_df.tcs_region==region
    results_2k15_df.loc[subset_mask_2k15,'region_rank']=range(1,sum(subset_mask_2k15)+1)
    subset_mask_2k16=results_2k16_df.tcs_region==region  #Works because both 2k15 and 2k16 have the same regions
    results_2k16_df.loc[subset_mask_2k16,'region_rank']=range(1,sum(subset_mask_2k16)+1)

print(sum(results_2k16_df.tcs_region=='Kolkata'))
results_2k16_df[results_2k16_df.tcs_region=='Kolkata'].tail()

#### Overall Percentile

results_2k15_df['overall_percentile']=1-results_2k15_df.Rank.rank(pct=True)
results_2k16_df['overall_percentile']=1-results_2k16_df.ranks.rank(pct=True)
results_2k15_df.loc[results_2k15_df.index[0],'overall_percentile']=1 #Manually setting percentile of topper to 100% (which might not be the case due to calculation errors)
results_2k16_df.loc[results_2k16_df.index[0],'overall_percentile']=1 #Manually setting percentile of topper to 100%
print(results_2k15_df.overall_percentile.head())
print(results_2k15_df.overall_percentile.tail())
print(results_2k16_df.overall_percentile.head())
print(results_2k16_df.overall_percentile.tail())

#### Zonal/Regional Percentile

for region in regions:
    subset_mask_2k15=results_2k15_df.tcs_region==region
    results_2k15_df.loc[subset_mask_2k15,'region_percentile']=1-results_2k15_df.loc[subset_mask_2k15,'region_rank'].rank(pct=True)
    subset_mask_2k16=results_2k16_df.tcs_region==region  #Works because both 2k15 and 2k16 have the same regions
    results_2k16_df.loc[subset_mask_2k16,'region_percentile']=1-results_2k16_df.loc[subset_mask_2k16,'region_rank'].rank(pct=True)

#Setting the regional toppers percentile to 100%
results_2k15_df.loc[results_2k15_df.region_rank==1,'region_percentile']=1
results_2k16_df.loc[results_2k16_df.region_rank==1,'region_percentile']=1

print(sum(results_2k16_df.tcs_region=='Ahmedabad'))
results_2k16_df[results_2k16_df.tcs_region=='Ahmedabad'].tail()

results_2k16_df[results_2k16_df.tcs_region=='Ahmedabad'].head()

results_2k15_df=results_2k15_df.rename(columns={'Rank':'overall_rank','users_name':'user_name'})

results_2k16_df=results_2k16_df.drop(['college_name_prerectified','tcs_region_prerectified'],axis=1).rename(columns={'ranks':'overall_rank'})

results_2k17_df=results_2k17_df.rename(columns={'round1_rank':'overall_rank','user1':'user_1','zonal_name':'tcs_region','zonal_rank':'region_rank'})

results_2k15_df.columns

results_2k16_df.columns

results_2k17_df.columns

results_df_combined_tableau=pd.DataFrame(results_2k15_df)
results_df_combined_tableau=results_df_combined_tableau.append(results_2k16_df)
results_df_combined_tableau=results_df_combined_tableau.append(results_2k17_df)

year_replace_dict={'2k15':2015,'2k16':2016,'2k17':2017}
results_df_combined_tableau.year=results_df_combined_tableau.year.apply(lambda x:year_replace_dict[x])
results_df_combined_tableau.year.dtype

results_df_combined_tableau.to_excel('./results_combined_for_tableau.xlsx',index=False)
